Fixed problems.

+++ uui-core/src/data/processing/views/Tree.ts
@@ -267,8 +267,8 @@ export class Tree<TItem, TId> {
                         ),
                     );
 
-                    if (existingItemParentId !== parentId) {
-                        const prevParentChildren = this.byParentId.get(existingItemParentId);
+                    if (existingItem && existingItemParentId !== parentId) {
+                        const prevParentChildren = this.byParentId.get(existingItemParentId) ?? [];
                         newByParentId.set(existingItemParentId, deleteFromList(id, prevParentChildren));
                     }
                 }
